research_stock_predictor/requirements.txt
# Research-Grade Multi-Modal Stock Prediction System Dependencies

# Data Handling
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0

# Deep Learning
tensorflow>=2.13.0

# NLP
textblob>=0.17.1

# Visualization & Deployment
streamlit>=1.29.0
plotly>=5.18.0
matplotlib>=3.8.0
seaborn>=0.13.0
